requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
httpx==0.28.1
python-dotenv==1.0.1
sse-starlette==2.2.1
groq==0.15.0
anthropic==0.42.0
openai==1.58.1
pydantic==2.10.4
pytest==8.3.4
ruff==0.8.6
